feat(telemetry): analytics always on in packaged builds, off in dev
